Reorganize test matchers
