--- src/backing/client/testing_util/README.md
@@ -1,12 +1,0 @@
-# `testing_util`
-
-Namespace containing various utilities for testing `Client` implementations.
-
-- `is_proto_equal` was taken from the [googleapis/google-cloud-cpp][]
-  repository ([source file][is-proto-equal]).
-- `fake_clock` was taken from the [googleapis/google-cloud-cpp][]
-  repository ([source file][fake-clock]).
-
-[googleapis/google-cloud-cpp]: https://github.com/googleapis/google-cloud-cpp
-[is-proto-equal]: https://github.com/googleapis/google-cloud-cpp/tree/a09f1358b3420fec9848629d2255a84e39182c45/google/cloud/testing_util
-[fake-clock]: https://github.com/googleapis/google-cloud-cpp/blob/master/google/cloud/spanner/testing/fake_clock.h
